Fix a few corner cases with itertools.groupby
This commit fixes a few corner cases that can be demonstrated by the
following test case:

```py
class TestGroupBy(unittest.TestCase):
    def test_unpacking(self):
        iterable = 'AAAAABBBBB'
        (_, a), (_, b) = groupby(iterable)
        self.assertEqual(list(a), [])
        self.assertEqual(list(b), [])

    def test_weird_iterating(self):
        g = groupby('AAAABBBBAAAAABBBBB')
        _, a = next(g)
        _, b = next(g)
        _, aa = next(g)
        self.assertEqual(list(a), [])
        self.assertEqual(list(b), [])
        self.assertEqual(list(aa), list('AAAAA'))
```

In the first case, `a` and `b` should both be empty lists because python
needs to call `next` on the `groupby` object 3 times to make sure that
the right number of items are yielded to be unpacked.  When the 3rd
`next` is called, the second group iterable (`b`) should be cleared.
However, it wasn't being cleared completely ...

The second case's desired behavior is a little more obvious (to me), but
prior to this change, we could have weirdness where if `a` was
used after `aa` had been created, `a` would still end up with elements
in it.

Both of these are corner cases that are well outside of the normal
use-cases of `groupby`, but this fix seems to be pretty simple (keep
track of which group `groupby` should be working on, any group which
isn't the current groupby group should raise StopIteration for any call
to `__next__`).